Refactor azure workspace test to dedicated class
